Move user_content_manager tests to a separate file
